+++ services/product/src/main/java/com/example/product/entity/Product.java
@@ -16,14 +16,10 @@ public class Product {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
-
+    private Long id;
     private String name;
-
     private String description;
-
     private BigDecimal price;
-
     private double availableQuantity;
 
     @ManyToOne
