+++ services/product/src/main/java/com/example/product/service/ProductService.java
@@ -19,60 +19,35 @@ public class ProductService {
     private final ProductRepository productRepository;
     private final CategoryRepository categoryRepository;
 
-    /**
-     * Create a new Product from ProductRequest DTO.
-     */
-    public Integer createProduct(ProductRequest request) {
+    public Long createProduct(ProductRequest request) {
         // Find category (throws error if not found)
         Category category = categoryRepository.findById(request.getCategoryId())
                 .orElseThrow(() -> new EntityNotFoundException("Category not found"));
-
-        // Map request → entity
         Product product = ProductMapper.toProduct(request, category);
-
-        // Save product
         productRepository.save(product);
-
-        // Return generated product ID
         return product.getId();
     }
 
-    /**
-     * Purchase products (reduce stock & return purchase details).
-     */
     public List<ProductPurchaseResponse> purchaseProducts(List<ProductPurchaseRequest> requests) {
         return requests.stream().map(req -> {
-            // Fetch product
             Product product = productRepository.findById(req.getProductId())
                     .orElseThrow(() -> new EntityNotFoundException("Product not found"));
-
-            // Check available stock
             if (product.getAvailableQuantity() < req.getQuantity()) {
                 throw new IllegalArgumentException("Not enough stock for product: " + product.getName());
             }
-
-            // Reduce stock
             product.setAvailableQuantity(product.getAvailableQuantity() - req.getQuantity());
             productRepository.save(product);
-
-            // Return response
             return ProductMapper.toPurchaseResponse(product, req.getQuantity());
         }).toList();
     }
 
-    /**
-     * Find product by ID.
-     */
-    public ProductResponse findById(Integer productId) {
+    public ProductResponse findById(Long productId) {
         Product product = productRepository.findById(productId)
                 .orElseThrow(() -> new EntityNotFoundException("Product not found"));
 
         return ProductMapper.toResponse(product);
     }
 
-    /**
-     * Find all products.
-     */
     public List<ProductResponse> findAll() {
         return productRepository.findAll()
                 .stream()
